% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Brexit}
\alias{Brexit}
\title{Articles on Brexit posted on the (former) Dutch news site nujij.nl.}
\format{
A data frame with eight variables:
\describe{
  \item{id}{Original article ID}
  \item{category}{Type of article (Dutch)}
  \item{pubDate}{Post date and time}
  \item{title}{Article heading}
  \item{text}{Article (body) text}
  \item{clicks}{Number of clicks received by the article}
  \item{votes}{Number of votes received by the article}
  \item{reactions}{Number of reactions posted on the article}
}
}
\usage{
Brexit
}
\description{
Articles on Brexit posted on the (former) Dutch news site nujij.nl.
}
\keyword{datasets}
