% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights4}
\alias{flights4}
\title{Subset of observations from the flights tibble in package nycflights13.}
\format{
A data frame with 20 variables:
\describe{
  \item{year, month, day}{Date of departure.}
  \item{dep_time, arr_time}{Actual departure and arrival times (format HHMM or HMM), local tz.}
  \item{sched_dep_time, sched_arr_time}{Scheduled departure and arrival times (format HHMM or HMM), local tz.}
  \item{dep_delay, arr_delay}{Departure and arrival delays, in minutes. Negative times represent early departures/arrivals.}
  \item{carrier}{Two letter carrier abbreviation. See airlines to get name.}
  \item{flight}{Flight number.}
  \item{tailnum}{Plane tail number. See planes for additional metadata.}
  \item{origin, dest}{Origin and destination. See airports for additional metadata.}
  \item{air_time}{Amount of time spent in the air, in minutes.}
  \item{distance}{Distance between airports, in miles.}
  \item{hour, minute}{Time of scheduled departure broken into hour and minutes.}
  \item{time_hour}{Scheduled date and hour of the flight as a POSIXct date. Along with origin, can be used to join flights data to weather data.}
  \item{arr_delay0}{Arrival delays without negative values.}
}
}
\usage{
flights4
}
\description{
Only flights originating from EWR and JFK have been selected.
Arrival delay of a flight without negative values has been added as variable arr_delay0.
}
\keyword{datasets}
