% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GlasgowFriends}
\alias{GlasgowFriends}
\title{Glasgow secondary school students and their (best) friends.}
\format{
A data frame with thirteen variables: 
\describe{
  \item{student}{respondent ID, as a character string}
  \item{neighbourhood}{student's home neighbourhood (fake)}
  \item{schooldist}{average distance from the neighbourhood to the school (in miles)}
  \item{hoodname}{(fake) name of the neighbourhood}
  \item{age}{respondent age, in years with one decimal digit}
  \item{sex}{respondent sex, boy or girl}
  \item{smoking_at_home}{any smokers at home, yes or no}
  \item{smoking_parents}{smoking by at least one parent, yes or no}
  \item{smoking_siblings}{smoking by at least one sibling, yes or no}
  \item{alcohol}{respondent alcohol consumption: 1 (none), 2 (once or twice a year), 3 (once a month), 4 (once a week) and 5 (more than once a week)}
  \item{cannabis}{respondent cannabis consumption: 1 (none), 2 (tried once), 3 (occasional) and 4 (regular)}
  \item{tobacco}{respondent tobacco consumption: 1 (none), 2 (occasional) and 3 (regular, i.e. more than once per week)}
  \item{money}{respondent's pocket money per month, in British pounds}
  \item{romantic}{whether the student had a romantic relation, yes or no}
  \item{wave}{time of observation, starting in February 1995, when the pupils were aged 13, and ending in January 1997}
  \item{friend_1}{code of a first student nominated as a friend by the respondent}
  \item{friend_2}{code of a second student nominated as a friend by the respondent}
  \item{friend_3}{code of a third student nominated as a friend by the respondent}
  \item{friend_4}{code of a fourth student nominated as a friend by the respondent}
  \item{friend_5}{code of a fifth student nominated as a friend by the respondent}
  \item{friend_6}{code of a sixth student nominated as a friend by the respondent}
  \item{bestfriend}{code of a student nominated as a/the best friend by the respondent}
  \item{bfperiod}{phase in the period of best friendship}
  \item{bfperiod}{wave for the phase in the period of best friendship}
}
}
\usage{
GlasgowFriends
}
\description{
The data derive from the same source as the Glasgow data. 
The data have been joined and simplified for practicing purposes.
A rough classification according of students according to their home 
neighbourhood was derived from the distances among students.
This data file is very untidy!
}
\keyword{datasets}
