% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glbwarm}
\alias{glbwarm}
\title{Opinions on global warming}
\format{
A data frame with 7 variables:
\describe{
  \item{govact}{Support for government action against climate change.}
  \item{posemot}{Positive emotions about climate change.}
  \item{negemot}{Negative emotions about climate change.}
  \item{ideology}{Respondent's political ideology (conservatism): Very Liberal, Liberal, Somewhat Liberal, Moderate; Middle of the Road, Somewhat Conservative, Conservative, Very Conservative.}
  \item{age}{Respondent's age.}
  \item{sex}{Respondent's sex.}
  \item{partyid}{Respondent's political party self-identification: democrat, independent, republican}
}
}
\source{
\url{http://www.afhayes.com/public/hayes2018data.zip}
}
\usage{
glbwarm
}
\description{
Data from a survey on attitudes towards global warming. The data set was collected by Erik Nisbet.
The data have been exported from the SPSS data set to CSV with value labels for categorical variables.
}
\keyword{datasets}
