% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surveyweek1}
\alias{surveyweek1}
\title{Subset of variables and observations from the first Friends & Families (MIT) weekly survey.}
\format{
A data frame with ten variables:
\describe{
  \item{participantID}{Original participant ID}
  \item{StartDate}{Date and time the participant started this survey}
  \item{EndDate}{Date and time the participant submitted this survey}
  \item{To the best of your recollection [...] March 8 - Sad or depressed}{Survey question on participant's mood}
  \item{To the best of your recollection [...] March 9 - Sad or depressed}{Survey question on participant's mood}
  \item{To the best of your recollection [...] March 10 - Sad or depressed}{Survey question on participant's mood}
  \item{To the best of your recollection [...] March 11 - Sad or depressed}{Survey question on participant's mood}
  \item{To the best of your recollection [...] March 12 - Sad or depressed}{Survey question on participant's mood}
  \item{To the best of your recollection [...] March 13 - Sad or depressed}{Survey question on participant's mood}
  \item{To the best of your recollection [...] March 14 - Sad or depressed}{Survey question on participant's mood}
}
}
\usage{
surveyweek1
}
\description{
Subset of variables and observations from the first Friends & Families (MIT) weekly survey.
}
\keyword{datasets}
