\name{UsingRTutorials}
\alias{UsingRTutorials}
\title{Tutorials for the course Using R for Data Wrangling, Analysis, and Visualization}
\usage{
learnr::run_tutorial(name, package = "UsingRTutorials")
}
\description{
Starts the tutorial named *name*.
}
\examples{
learnr::run_tutorial("introduction", package = "UsingRTutorials")
}
