% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsingRTutorials.R
\docType{package}
\name{UsingRTutorials}
\alias{UsingRTutorials}
\title{UsingRTutorials: Tutorials for the course Using R for Data Wrangling, Analysis, and Visualization}
\usage{
learnr::run_tutorial(name, package = "UsingRTutorials")

Starts the tutorial named *name*.
}
\description{
The UsingRTutorials package provides an interactive tutorial for each session
in the R for Data Wrangling, Analysis, and Visualization course at the
Unversity of Amsterdam, Department of Communication Science.
}
\examples{

\dontrun{
  learnr::run_tutorial("introduction", package = "UsingRTutorials")
 }
}
